Rearrange routes, add search routes, but not functional yet
